requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
arrow==1.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
discord==2.3.2
discord-ui==5.1.6
discord.py==2.3.2
frozenlist==1.4.1
ics==0.7.2
idna==3.7
multidict==6.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
ratelimit==2.2.1
requests==2.31.0
six==1.16.0
TatSu==5.12.1
types-python-dateutil==2.9.0.20240316
urllib3==2.2.1
yarl==1.9.4
zope.interface==6.3
google-api-python-client==2.143.0